Fix navigation/safe arg Tests
