+++ app/build.gradle
@@ -18,9 +18,6 @@ android {
         vectorDrawables.useSupportLibrary = true
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        // Reset state between tests
-        //testInstrumentationRunnerArguments clearPackageData: "true" TODO("ask about use w/ running multiple nav tests as opposed to w/ room")
-        // TODO("Ask about disabling in certain tests")
     }
 
     buildFeatures {
@@ -42,11 +39,6 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
-
-    testOptions {
-        // Reset state between tests
-        //execution = "ANDROIDX_TEST_ORCHESTRATOR"
-    }
 }
 
 dependencies {
@@ -63,7 +55,8 @@ dependencies {
     androidTestImplementation "androidx.test.ext:junit:$junit_androidx_ext_version"
     androidTestImplementation "androidx.test.espresso:espresso-core:$espresso_version"
     debugImplementation "androidx.fragment:fragment-testing:$fragment_version"
-    androidTestImplementation "com.android.support.test.espresso:espresso-contrib:$espresso_contrib_version" // Nav
+    androidTestImplementation "com.android.support.test.espresso:espresso-contrib:$espresso_contrib_version"
+    // Nav
     androidTestImplementation "androidx.navigation:navigation-testing:$nav_version"
 
     // LiveData
